Deleting heat stack when deleting assembly
- Setting assembly status to DELETING
- Asynchronously calling heat stack delete
- Destroying the assembly object only if heat
  stack successfully got deleted.
  If stack delete errors out, setting the
  assembly status to ERROR_HEAT_STACK_DELETE
- Added sleep to functionaltests to allows for
  children assemblies to be deleted before
  delete_plan is called.

Notes:
1) If Heat stack-delete errors out, a stack with
   name = assembly_name will exist in the system
   (most likely in an indeterminate state). If
   Solum receives a subsequent call to create a
   new assembly with the same assembly name then
   Heat stack-creation most likely would error out
   because of the clashing stack name.

   One approach to solve this issue will be to append
   assembly_id to the assembly_name when creating a
   heat stack

   Created bug for this: #1316304

2) Once we start storing stack_id in the component
   table we can avoid the call to heat to get
   the stack id.

Partial-bug: #1314626
 (Helped with testing functional test changes)

+++ functionaltests/api/v1/public/test_trigger.py
@@ -14,6 +14,7 @@ import json
 import requests
 
 from functionaltests.api import base
+from tempest import exceptions as tempest_exceptions
 
 assembly_data = {'name': 'test_assembly',
                  'description': 'desc assembly'}
@@ -59,7 +60,11 @@ class TestTriggerController(base.TestCase):
     def _delete_assembly(self, assembly_uuid, plan_uuid):
         resp, body = self.client.delete('v1/assemblies/%s' % assembly_uuid)
         self.assertEqual(resp.status, 204)
-        self._delete_plan(plan_uuid)
+
+        if self.client.assembly_delete_done(assembly_uuid):
+            self._delete_plan(plan_uuid)
+        else:
+            self.assertRaises(tempest_exceptions.TimeoutException)
 
     def _delete_plan(self, plan_uuid):
         resp, body = self.client.delete('v1/plans/%s' % plan_uuid)
