+++ functionaltests/api/v1/test_assembly.py
@@ -45,7 +45,11 @@ class TestAssemblyController(base.TestCase):
     def _delete_assembly(self, uuid, plan_uuid):
         resp, _ = self.client.delete('v1/assemblies/%s' % uuid)
         self.assertEqual(resp.status, 204)
-        self._delete_plan(plan_uuid)
+
+        if self.client.assembly_delete_done(uuid):
+            self._delete_plan(plan_uuid)
+        else:
+            self.assertRaises(tempest_exceptions.TimeoutException)
 
     def _delete_plan(self, plan_uuid):
         resp, _ = self.client.delete('v1/plans/%s' % plan_uuid)
@@ -146,7 +150,11 @@ class TestAssemblyController(base.TestCase):
         resp, body = self.client.delete('v1/assemblies/%s' % uuid)
         self.assertEqual(resp.status, 204)
         self.assertEqual(body, '')
-        self._delete_plan(plan_uuid)
+
+        if self.client.assembly_delete_done(uuid):
+            self._delete_plan(plan_uuid)
+        else:
+            self.assertRaises(tempest_exceptions.TimeoutException)
 
     def test_assemblies_delete_not_found(self):
         self.assertRaises(tempest_exceptions.NotFound,
