+++ functionaltests/api/v1/test_component.py
@@ -46,7 +46,11 @@ class TestComponentController(base.TestCase):
     def _delete_assembly(self, assembly_uuid, plan_uuid):
         resp, body = self.client.delete('v1/assemblies/%s' % assembly_uuid)
         self.assertEqual(resp.status, 204)
-        self._delete_plan(plan_uuid)
+
+        if self.client.assembly_delete_done(assembly_uuid):
+            self._delete_plan(plan_uuid)
+        else:
+            self.assertRaises(tempest_exceptions.TimeoutException)
 
     def _delete_plan(self, plan_uuid):
         resp, body = self.client.delete('v1/plans/%s' % plan_uuid)
